Use TableBlockSdk to change table name
