+++ apps/web/modules/components/editor/blocks/sdk/index.ts
@@ -1,0 +1,1 @@
+export * as TableBlockSdk from './table';
