+++ apps/web/modules/components/editor/blocks/table/table-block-type-picker.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import { Input } from '~/modules/design-system/input';
 import { Text } from '~/modules/design-system/text';
 import { SelectedEntityType } from '~/modules/entity';
-import { ResultItem, ResultsList } from '../../../autocomplete/results-list';
+import { ResultItem, ResultsList } from '../../../entity/autocomplete/results-list';
 import { Spacer } from '~/modules/design-system/spacer';
 import { useTypesStore } from '~/modules/type/types-store';
