--- apps/web/modules/components/entity/editor/blocks/sdk/table.ts
@@ -1,11 +1,0 @@
-export function createTableBlock() {
-  return;
-}
-
-export function editTableBlock() {
-  return;
-}
-
-export function deleteTableBlock() {
-  return;
-}
